requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
supabase
torch>=2.0.0
torch-geometric>=2.3.0
numpy>=1.24.0
matplotlib==3.8.2
scikit-learn>=1.3.0
pandas>=2.0.0
python-multipart==0.0.6
pydantic==2.5.2
httpx
tqdm>=4.65.0
networkx>=3.1
sentence-transformers>=2.2.0
PyJWT[crypto]
cryptography
pydantic-settings 